refactor: moved most of the code to internal package
